// frontend/src/pages/Admin/ManageReferenceProject/types.ts
export interface SectionDataFromSelection {
  endDate: string
  id: string
  projectName: string
  sectionId: string
  sectionName: string
  startDate: string
}

export interface InputValues {
  _id: string
  description: string
  projectName: string
  projectUrl: string
  sectionId: string
  sectionName: string
}